Add check for non-existing table internal name for delete table
This change added check for non-existing internal name for tables
which caused exception when trying to delete table in CREATE_FAILED
or DELETE_FALED status if table internal name does not exist.

Closes-bug: #1422908

+++ magnetodb/storage/driver/cassandra/cassandra_impl.py
@@ -230,6 +230,12 @@ class CassandraStorageDriver(driver.StorageDriver):
         :raises: BackendInteractionException
         """
 
+        if not table_info.internal_name:
+            LOG.info(("Table '{}' with tenant id '{}', id '{}' does not have "
+                      "valid internal name. Unable or no need to delete.").
+                     format(table_info.name, context.tenant, table_info.id))
+            return
+
         query = 'DROP TABLE IF EXISTS ' + table_info.internal_name
 
         self.__cluster_handler.execute_query(query)
